Correct fix! Was for when run unit test 1st time and no directory was there.

+++ core/src/test/java/org/bitcoinj/core/LevelDbFullPrunedBlockChainTest.java
@@ -25,7 +25,7 @@ public class LevelDbFullPrunedBlockChainTest extends AbstractFullPrunedBlockChai
 
     private void deleteFiles() {
         File f=new File("test-leveldb");
-        if(f!=null){
+        if(f!=null && f.exists()){
 	        for (File c : f.listFiles())
 	        	c.delete();
         }
